requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-dotenv==0.19.1
openai==0.27.0
starlette==0.16.0
python-decouple==3.8
itsdangerous==2.1.2
json
